# src/main/resources/functional_requirements/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/ui/payment/{id}": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaymentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "updatePayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "transition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "deletePayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/ui/settlement-quote": {
      "get": {
        "tags": [
          "settlement-quote-controller"
        ],
        "operationId": "getAllSettlementQuotes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataSettlementQuote"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settlement-quote-controller"
        ],
        "operationId": "createSettlementQuote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettlementQuote"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataSettlementQuote"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/payment-file": {
      "get": {
        "tags": [
          "payment-file-controller"
        ],
        "operationId": "getAllPaymentFiles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataPaymentFile"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "payment-file-controller"
        ],
        "operationId": "createPaymentFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPaymentFile"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/settlement-quote/{id}": {
      "get": {
        "tags": [
          "settlement-quote-controller"
        ],
        "operationId": "getSettlementQuoteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataSettlementQuote"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/payment/business/{paymentId}": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaymentByBusinessId",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      }
    },
    "/ui/payment-file/{id}": {
      "get": {
        "tags": [
          "payment-file-controller"
        ],
        "operationId": "getPaymentFileById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPaymentFile"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Payment": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "valueDate": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string"
          },
          "sourceReference": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "loanId": {
            "type": "string"
          },
          "matchingReference": {
            "type": "string"
          },
          "virtualAccountReference": {
            "type": "string"
          },
          "allocation": {
            "$ref": "#/components/schemas/PaymentAllocation"
          },
          "paymentFileId": {
            "type": "string"
          },
          "sourceType": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "capturedBy": {
            "type": "string"
          },
          "matchedAt": {
            "type": "string",
            "format": "date-time"
          },
          "allocatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "PaymentAllocation": {
        "type": "object",
        "properties": {
          "interestAmount": {
            "type": "number"
          },
          "feesAmount": {
            "type": "number"
          },
          "principalAmount": {
            "type": "number"
          },
          "totalAllocated": {
            "type": "number"
          },
          "unallocatedAmount": {
            "type": "number"
          },
          "allocationNotes": {
            "type": "string"
          }
        }
      },
      "EntityMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "state": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataPayment": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Payment"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
      "ModelSpec": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SettlementBreakdown": {
        "type": "object",
        "properties": {
          "principalAsOf": {
            "type": "number"
          },
          "interestToDate": {
            "type": "number"
          },
          "feeCalculation": {
            "type": "number"
          },
          "feePolicy": {
            "type": "string"
          },
          "feeRate": {
            "type": "number"
          },
          "calculationNotes": {
            "type": "string"
          }
        }
      },
      "SettlementQuote": {
        "type": "object",
        "properties": {
          "settlementQuoteId": {
            "type": "string"
          },
          "loanId": {
            "type": "string"
          },
          "asOfDate": {
            "type": "string",
            "format": "date"
          },
          "expiryDate": {
            "type": "string",
            "format": "date"
          },
          "quotedDate": {
            "type": "string",
            "format": "date"
          },
          "outstandingPrincipal": {
            "type": "number"
          },
          "accruedInterest": {
            "type": "number"
          },
          "settlementFee": {
            "type": "number"
          },
          "totalPayoffAmount": {
            "type": "number"
          },
          "breakdown": {
            "$ref": "#/components/schemas/SettlementBreakdown"
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedBy": {
            "type": "string"
          },
          "paymentId": {
            "type": "string"
          },
          "executedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataSettlementQuote": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/SettlementQuote"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
...
      "FileValidationError": {
        "type": "object",
        "properties": {
          "lineNumber": {
            "type": "integer",
            "format": "int32"
          },
          "errorCode": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "fieldName": {
            "type": "string"
          },
          "fieldValue": {
            "type": "string"
          }
        }
      },
      "PaymentFile": {
        "type": "object",
        "properties": {
          "paymentFileId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "fileHash": {
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "validRecords": {
            "type": "integer",
            "format": "int32"
          },
          "invalidRecords": {
            "type": "integer",
            "format": "int32"
          },
          "processedRecords": {
            "type": "integer",
            "format": "int32"
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileValidationError"
            }
          },
          "createdPaymentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "validationStartedAt": {
            "type": "string",
            "format": "date-time"
          },
          "validationCompletedAt": {
            "type": "string",
            "format": "date-time"
          },
          "importStartedAt": {
            "type": "string",
            "format": "date-time"
          },
          "importCompletedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataPaymentFile": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/PaymentFile"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
...
    }
  }
}